Reorganize things.
