+++ app/routes/community/index.tsx
@@ -3,7 +3,7 @@ import { useLoaderData } from "@remix-run/react"
 import { LoaderFunction } from "@remix-run/server-runtime"
 import { fetchLatestTopics } from "~/cms/utils/fetch-discourse-api"
 import { fetchFlips } from "~/cms/utils/fetch-flips"
-import { allEvents } from "~/component-data/Events"
+import { allEvents } from "~/ui/component-data/Events"
 import { getMetaTitle } from "~/root"
 import { CommunityMembersProps } from "~/ui/design-system/src/lib/Components/CommunityMembers"
 import { Default as DefaultCommunityMembers } from "~/ui/design-system/src/lib/Components/CommunityMembers/CommunityMembers.stories"
