+++ app/routes/concepts/index.tsx
@@ -1,7 +1,7 @@
 import { LoaderFunction, MetaFunction } from "@remix-run/node"
 import { getMetaTitle } from "~/root"
 import { temporarilyRedirectToComingSoon } from "~/utils/features"
-import { ConceptsPage } from "../../ui/design-system/src/lib/Pages/ConceptsPage"
+import { ConceptsPage } from "~/ui/design-system/src/lib/Pages/ConceptsPage"
 import { data } from "~/cms/route-presets/concepts/data"
 
 export const meta: MetaFunction = () => ({
