+++ app/routes/tools/index.tsx
@@ -1,6 +1,6 @@
 import { MetaFunction } from "@remix-run/node"
 import { getMetaTitle } from "~/root"
-import ToolsPage from "../../ui/design-system/src/lib/Pages/ToolsPage"
+import ToolsPage from "~/ui/design-system/src/lib/Pages/ToolsPage"
 import { data } from "../../cms/route-presets/tools/data"
 
 export const meta: MetaFunction = () => ({
