+++ app/ui/component-data/NavigationBar/navCardsData.tsx
