+++ app/ui/component-data/NavigationBar/networkTabData.tsx
