+++ app/ui/component-data/Tools/index.tsx
@@ -1,11 +1,11 @@
 import { SDKCardProps } from "~/ui/design-system/src/lib/Components/SDKCard"
-import ToolCliIconSrc from "../../ui/design-system/images/tools/tool-cli.svg"
-import ToolEmulatorIconSrc from "../../ui/design-system/images/tools/tool-emulator.svg"
-import ToolFclIconSrc from "../../ui/design-system/images/tools/tool-fcl.svg"
-import ToolTestingIconSrc from "../../ui/design-system/images/tools/tool-testing.svg"
-import ToolVsCodeIconSrc from "../../ui/design-system/images/tools/tool-vscode.svg"
-import ToolPortIconSrc from "../../ui/design-system/images/tools/tool-port.svg"
-import CodeIcon from "../../ui/design-system/images/tools/tool-default.svg"
+import ToolCliIconSrc from "~/ui/design-system/images/tools/tool-cli.svg"
+import ToolEmulatorIconSrc from "~/ui/design-system/images/tools/tool-emulator.svg"
+import ToolFclIconSrc from "~/ui/design-system/images/tools/tool-fcl.svg"
+import ToolTestingIconSrc from "~/ui/design-system/images/tools/tool-testing.svg"
+import ToolVsCodeIconSrc from "~/ui/design-system/images/tools/tool-vscode.svg"
+import ToolPortIconSrc from "~/ui/design-system/images/tools/tool-port.svg"
+import CodeIcon from "~/ui/design-system/images/tools/tool-default.svg"
 
 // Flow Dev Tools
 const cliTool: SDKCardProps = {
