+++ app/ui/design-system/src/lib/Components/LandingPageLinks/GithubLinks.tsx
@@ -11,7 +11,7 @@ import {} from "@headlessui/react"
 import clsx from "clsx"
 import debounce from "lodash.debounce"
 import { useCallback, useEffect, useMemo, useRef, useState } from "react"
-import { GITHUB_URL } from "../../../../../../cms/constants"
+import { GITHUB_URL } from "../../../../../../constants"
 import { ReactComponent as ExternalLinkIcon } from "../../../../images/content/external-link"
 import { ReactComponent as DropdownArrowIcon } from "../../../../images/misc/dropdown-arrow"
 import { ReactComponent as GithubIcon } from "../../../../images/social/github"
